requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pymongo==4.6.1
motor==3.3.2
faster-whisper==1.0.3
gtts==2.5.1
pydub==0.25.1
httpx==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
scikit-learn==1.4.0
redis==7.4.0
av==17.0.0
